Soft Delete için Unique anatasyonu kaldırdım. Kontrolleri servise ekledim.

+++ qr-tracking-system-dss-be/src/main/java/com/mk/qr_tracking_system_dss/service/Impl/UserServiceImpl.java
@@ -29,6 +29,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(Long userId) {
+        User currentUser = getCurrentUser();
+        if (currentUser.getId().equals(userId)) {
+            throw new IllegalArgumentException("Kendinizi silemezsiniz.");
+        }
         userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("Bu ID ile kullanıcı bulunamadı."));
         userRepository.deleteById(userId);
